feature: add option to download iris data

+++ hooks/post_gen_project.py
@@ -7,10 +7,16 @@ def main() -> None:
         "yes",
         "True",
     ]
+    include_iris_data = "{{ cookiecutter.include_iris_data }}".strip() in [
+        "yes",
+        "True",
+    ]
 
     if not include_iris_example:
         shutil.rmtree("src/{{ cookiecutter.python_package }}/pipelines/iris_example")
         os.remove("conf/base/parameters/iris_example.yml")
+
+    if not include_iris_example and not include_iris_data:
         os.remove("data/01_raw/iris.csv")
